+++ src/SQLProvider/SqlRuntime.Linq.fs
@@ -367,8 +367,25 @@ module internal QueryImplementation =
                     | SqlNegativeCondOp(op,OptionalConvertOrTypeAs(SqlColumnGet(ti,key,_)),OptionalConvertOrTypeAs(OptionalFSharpOptionValue((((:? MemberExpression) | (:? MethodCallExpression) | (:? NewExpression)) as meth))))
                     | SqlCondOp(op,OptionalConvertOrTypeAs(OptionalFSharpOptionValue((((:? MemberExpression) | (:? MethodCallExpression) | (:? NewExpression)) as meth))),OptionalConvertOrTypeAs(SqlColumnGet(ti,key,_)))
                     | SqlNegativeCondOp(op,OptionalConvertOrTypeAs(OptionalFSharpOptionValue((((:? MemberExpression) | (:? MethodCallExpression) | (:? NewExpression)) as meth))),OptionalConvertOrTypeAs(SqlColumnGet(ti,key,_))) ->
+
                         paramNames.Add(ti) |> ignore
-                        Some(ti,key,op,Some(Expression.Lambda(meth).Compile().DynamicInvoke()))
+                        let invokedResult = Expression.Lambda(meth).Compile().DynamicInvoke()
+
+                        let handleNullCompare() =
+                            match op with
+                            | ConditionOperator.Equal -> Some(ti,key,ConditionOperator.IsNull,None)
+                            | ConditionOperator.NotEqual -> Some(ti,key,ConditionOperator.NotNull,None)
+                            | _ -> Some(ti,key,op,Some(invokedResult))
+
+                        if invokedResult = null then handleNullCompare()
+                        else
+                        let retType = invokedResult.GetType()
+                        if retType.IsGenericType && retType.FullName.StartsWith("Microsoft.FSharp.Core.FSharpOption") then
+                            let gotVal = retType.GetProperty("Value") // Option type Some should not be SQL-parameter.
+                            match gotVal.GetValue(invokedResult, [||]) with
+                            | null -> handleNullCompare()
+                            | r -> Some(ti,key,op,Some(r))
+                        else Some(ti,key,op,Some(invokedResult))
                     | SqlColumnGet(ti,key,ret) when exp.Type.FullName = "System.Boolean" -> 
                         paramNames.Add(ti) |> ignore
                         Some(ti,key,ConditionOperator.Equal, Some(true |> box))
@@ -569,9 +586,25 @@ module internal QueryImplementation =
                                     ForeignTable = {Schema="";Name="";Type=""};
                                     OuterJoin = false; RelDirection = RelationshipDirection.Parents }
                     SelectMany(sourceAlias,destAlias,LinkQuery(data),outExp)
+                | OptionalConvertOrTypeAs(MethodCall(Some(Lambda([_],MethodCall(_,MethodWithName "CreateEntities",[String destEntity]))),(MethodWithName "Invoke"),_)) ->
+                    let sourceAlias =
+                        if source.TupleIndex.Contains projectionParams.[0].Name then projectionParams.[0].Name
+                        else
+                        match source.SqlExpression with
+                        | BaseTable(a, t) -> t.Name
+                        | _ -> projectionParams.[0].Name
+                    let table = Table.FromFullName destEntity
+                    let destAlias = table.Name
+
+                    if source.TupleIndex.Any(fun v -> v = sourceAlias) |> not then source.TupleIndex.Add(sourceAlias)
+                    if source.TupleIndex.Any(fun v -> v = destAlias) |> not then source.TupleIndex.Add(destAlias)
+                    SelectMany(sourceAlias,destAlias,CrossJoin(table.Name,table),outExp)
+                | MethodCall(None, (MethodWithName "AsQueryable"), [innerExp]) ->
+                    processSelectManys projectionParams.[0].Name innerExp outExp projectionParams source
+                | MethodCall(None, (MethodWithName "Select"), [ createRelated; OptionalQuote (Lambda([ v1 ], _) as lambda) ]) as whole ->
+                    let outExp = processSelectManys projectionParams.[0].Name createRelated outExp projectionParams source
+                    Projection(whole,outExp)
                 | _ -> failwith ("Unknown: " + inExp.ToString())
-
-
 
              // Possible Linq method overrides are available here: 
              // https://referencesource.microsoft.com/#System.Core/System/Linq/IQueryable.cs
@@ -592,7 +625,7 @@ module internal QueryImplementation =
                     | MethodCall(None, (MethodWithName "OrderBy" | MethodWithName "OrderByDescending" as meth), [SourceWithQueryData source; OptionalQuote (Lambda([ParamName param], OptionalConvertOrTypeAs (SqlColumnGet(entity,key,_)))) ]) ->
                         let alias =
                              match entity with
-                             | "" when source.SqlExpression.HasAutoTupled() -> param
+                             | "" when source.SqlExpression.HasAutoTupled() && source.TupleIndex.Contains param -> param
                              | "" -> ""
                              | _ -> Utilities.resolveTuplePropertyName entity source.TupleIndex
                         let ascending = meth.Name = "OrderBy"
@@ -611,7 +644,7 @@ module internal QueryImplementation =
                     | MethodCall(None, (MethodWithName "ThenBy" | MethodWithName "ThenByDescending" as meth), [SourceWithQueryData source; OptionalQuote (Lambda([ParamName param], OptionalConvertOrTypeAs (SqlColumnGet(entity,key,_)))) ]) ->
                         let alias =
                             match entity with
-                            | "" when source.SqlExpression.HasAutoTupled() -> param
+                            | "" when source.SqlExpression.HasAutoTupled() && source.TupleIndex.Contains param -> param
                             | "" -> ""
                             | _ -> Utilities.resolveTuplePropertyName entity source.TupleIndex
                         let ty = typedefof<SqlOrderedQueryable<_>>.MakeGenericType(meth.GetGenericArguments().[0])
@@ -854,7 +887,7 @@ module internal QueryImplementation =
 
                         let alias =
                              match entity with
-                             | "" when source.SqlExpression.HasAutoTupled() -> param
+                             | "" when source.SqlExpression.HasAutoTupled() && source.TupleIndex.Contains param -> param
                              | "" -> ""
                              | _ -> resolveTuplePropertyName entity source.TupleIndex