+++ RELEASE_NOTES.md
@@ -1,3 +1,18 @@
+### 1.1.22 - 09.11.2017
+* Option type fix for exactlyOneOrDefault. PR #491
+
+### 1.1.21 - 07.11.2017
+* Initial support for cross joins. PR #489
+
+### 1.1.20 - 06.11.2017
+* Option type exactlyOneOrDefault fix PR #484
+* Oracle join columns fix PR #455
+
+### 1.1.19 - 01.11.2017
+* Oracle column name resolution PR #480 
+* Procedure name overload fix PR #476 
+* Allow option type return in where-lambda. #478 
+
 ### 1.1.18 - 26.10.2017
 * Previous type provider SDK returned #472 #470 #468 #466 #465
 * More info to tooltips.